The COBOL source code that converts time to words
Two files are uploaded. The source code in the file TimeToWords-CurrentDate.cbl takes the current time using FUNCTION CURRENT-DATE and converts that to words. The source code in the file TimeToWords-InputFile.cbl reads an input file that contains 1440 records, each record having the time in HHMM format, and writes the corresponding words to an output file.